requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
langgraph==0.2.55
langchain-core==0.3.25
langchain-google-genai==2.0.7
chromadb==0.5.23
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
